move new strategies
